companyapi/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
dj-database-url==2.0.0
Django==4.2
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-heroku==0.3.1
django-js-asset==2.0.0
django-multiselectfield==0.1.12
django-storages==1.13.2
djangorestframework==3.14.0
gcloud==0.17.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
jws==0.1.3
mysqlclient==2.1.1
oauth2client==3.0.0
Pillow==9.5.0
protobuf==4.23.2
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.4.3
pyparsing==3.0.9
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2023.3
requests==2.31.0
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.6.3
urllib3==1.26.7
whitenoise==6.5.0
djangorestframework==x.x.x
